Renamed to ReportService and lowered case of CSS folder
