+++ src/components/AdditionalInformation.jsx
@@ -4,7 +4,7 @@ import ErrorMsg from "./ErrorMessage";
 import FormContainer from './FormContainer';
 import { IsFormComplete } from "../utilities/FormHelpers";
 import ButtonDisplay from "./buttonDisplay";
-import submitReport from "../utilities/submitReport";
+import { SubmitReport} from "../utilities/ReportService";
 import { GoHome, GoBack } from '../Routing';
 
 const AdditionalInformation = props => {
@@ -17,7 +17,7 @@ const AdditionalInformation = props => {
 	}
 
 	const SubmitTheForm = (clickEvent) => {
-		 submitReport(clickEvent, props);
+		SubmitReport(clickEvent, props);
 	}
 
 	const callPreviousForm = () => {
